% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{aggregate_price_and_mean_windows}
\alias{aggregate_price_and_mean_windows}
\title{Get Price and historic price mean windows from price relatives}
\usage{
aggregate_price_and_mean_windows(
  price_relative_matrix,
  decay_factor,
  price_window_size,
  historic_mean_window_size,
  initial_prices = NULL,
  include_missing = FALSE,
  .check_input = TRUE
)
}
\arguments{
\item{price_relative_matrix}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{p_{t+1} / p_t}, i.e.
the ratio of trading price to next price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{decay_factor}{the decay factor for the mean (see description)}

\item{historic_mean_window_size}{The size of the historic price mean
window to include}

\item{initial_prices}{A vector holding the initial prices of
each asset \eqn{i}}

\item{include_missing}{If \code{TRUE}, includes rows for trading
periods with either missing windows or missing next values/momenta}

\item{.check_input}{If \code{TRUE}, checks that input meets the
given assumptions.}

\item{momentum_window_size}{determines how momentum is computed.}
}
\value{
A tibble (arranged by \code{c(trading_period, asset)})
  subclassed to be of class \code{@eval{AGG_WINDOW_CLASS_NAME}}
  with columns
  \describe{
  \item{trading_period}{The trading period during which prices change}
  \item{asset}{The name (or column index if no names are present)
               of the asset, corresponding with the column names
               of the argument \code{price_relative_matrix}}
  \item{price}{The price before the prices change at the end of
               the trading period}
  \item{next_price}{The price after the prices change at the end of
                    the trading period}
  \item{price_window}{The \code{price_window_size} previous prices
                      (including the price during the trading period)
                      in chronological order (i.e. oldest first,
                      newest last)}
  \item{historic_price_mean}{The historic price mean
                             with decay factor \code{decay_factor}
                             (see \link{compute_historic_price_means})
                             during trading (before prices changes at the end
                             of the current trading period)}
  \item{next_historic_price_mean}{The historic price mean
                                  with decay factor \code{decay_factor}
                                 (see \link{compute_historic_price_means})
                                 after trading (after prices change
                                  at the end of the current trading period)}
  \item{historic_price_mean_window}{The previous
                                    \code{historic_mean_window_size}
                                    many historic price means. These
                                    are not used to compute anything,
                                    but may be desired as features.
                                    They are also in chronological order,
                                    oldest -> newest.}
 }
 You should not modify the values of these columns except using
 methods defined in this package (unless you know what you're doing...
 you know who you are)
}
\description{
Given price relatives and window sizes, returns a tibble
with price windows, historic mean windows, and momentum
values if desired.
}
