% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{aggregated_windows_var_to_matrix}
\alias{aggregated_windows_var_to_matrix}
\title{Puts variable in trading_period x asset matrix}
\usage{
aggregated_windows_var_to_matrix(agg_windows, variable_name)
}
\arguments{
\item{agg_windows}{a \code{@eval{AGG_WINDOW_CLASS_NAME} } instance
as returned from \code{\link{aggregate_price_and_mean_windows}}.}

\item{variable_name}{A string which is a variable name of a double
vector column of \code{agg_windows}.}
}
\value{
A matrix with entries from variable \code{variable_name},
    the column names being the \code{asset} variable from \code{agg_windows}
    and the row names being the \code{trading_period} variable from
    \code{agg_windows}.
}
\description{
Puts variable in trading_period x asset matrix
}
