% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_best_stock}
\alias{backtest_best_stock}
\title{backtests a best stock strategy}
\usage{
backtest_best_stock(price_relative_matrix, transaction_rate)
}
\description{
Backtests a stragey which buys the best stock and sits on it
for the duration. This is a hindsight strategy.
}
