% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_buyandhold}
\alias{backtest_buyandhold}
\title{backtest a buy and hold strategy}
\usage{
backtest_buyandhold(price_relative_matrix, transaction_rate, initial_portfolio)
}
\arguments{
\item{initial_portfolio}{a vector whose \eqn{i}th entry is the
amount of wealth in the \eqn{i}th entry. Should
be a portfolio for the assets in \code{price_relatives}.
This is the portfolio before the first trading period.}
}
\description{
backtest a buy and hold strategy
}
