% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_exponential_gradient}
\alias{backtest_exponential_gradient}
\title{Backtests using a exponential gradient rule}
\usage{
backtest_exponential_gradient(
  price_relative_matrix,
  transaction_rate,
  learning_rate
)
}
\arguments{
\item{price_relative_matrix}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{p_{t+1} / p_t}, i.e.
the ratio of trading price to next price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{transaction_rate}{The percentage of each transaction (buy and sell)
spent on broker fees}

\item{learning_rate}{The learning rate (\eqn{\eta} in the referenced
paper), a higher learning rate corresponds to a higher
sensitivitiy as one would expect.}
}
\value{
a matrix with the same number of columns and one more
row than \code{price_relative_matrix}, row \eqn{i}
is the portfolio after a trade during period \eqn{i}, i.e.
right before the \eqn{i}th price relatives change the prices.
}
\description{
backtests using an exponential gradient rule.
See this paper
\url{http://rob.schapire.net/papers/HelmboldScSiWa98.pdf}
("Online Portfolio Selection Using Multiplicative Updates"
by David P Helmholt, Robert E. Schapire and Yoram Singer,
and Manfred K. Warmuth)
}
\details{
has a multiplicative weight update rule, and initializes to
uniform portfolio
}
