% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_market}
\alias{backtest_market}
\title{backtests a market (uniform buy and hold) strategy}
\usage{
backtest_market(price_relative_matrix, transaction_rate)
}
\description{
Backtests a strategy which buys an equal amount of all stocks
and then holds them.
}
