% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{backtest_portfolio_selector}
\alias{backtest_portfolio_selector}
\title{backtest to get cumulative wealth and portfolios over time}
\usage{
backtest_portfolio_selector(
  pf_select,
  first_trading_period,
  last_trading_period
)
}
\arguments{
\item{pf_select}{the \link[=new_pfselectstrat]{pfselectstrat} instance to use}

\item{first_trading_period}{the trading period to start backtesting
on (right before prices change)}

\item{last_trading_period}{(OPTIONAL) if supplied, only backtests
up through \code{max_trading_period} (instead of through all
available trading periods). This is the last price change we experience}
}
\value{
A data frame which has
    @format{
      @describe{
         \item{trading_period}{The row index into \code{pf_select}'s
             price_relatives, indicating the current trading period}
         \item{cumulative_wealth}{The factor of wealth increase accumulated
             up to this trading period}
         \item{<Asset Names>}{Names matching the column names of
             \code{pf_select@price_relatives} indicating the
             portion of total wealth in that asset at the given trading
             period}
      }
    }
}
\description{
Run a portfolio selection strategy from the given trading
period until the end of its data, reporting the cumulative
wealth at each stage and the portfolio before each
trading period
}
