% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{best_stock}
\alias{best_stock}
\title{Makes a new \code{best_stock} strategy}
\usage{
best_stock(price_relatives, last_trading_period)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the current
price divided by the previous price.}

\item{last_trading_period}{the last trading period after
which we "sell" our stocks (we need this to determine
which stock is the best for the \code{best_stock} strategy).}
}
\value{
a new \code{buyandhold} object with the given price relatives
}
\description{
Makes a \code{buyandhold} strategy from the given price
relatives, which is a subclass of \code{pfselectstrat}.
}
\note{
if last_trading_period is missing, it defaults
     to the last available trading period
}
\seealso{
new_best_stock

Other buyandhold: 
\code{\link{new_buyandhold}()},
\code{\link{uniform_bah}()},
\code{\link{validate_buyandhold}()}
}
\concept{buyandhold}
