% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{buyandhold}
\alias{buyandhold}
\title{Makes a new \code{buyandhold} strategy}
\usage{
buyandhold(price_relatives, transaction_rate)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}
}
\value{
a new \code{buyandhold} object with the given price relatives
}
\description{
Makes a \code{\link[=new_buyandhold]{buyandhold}} strategy from the given price
relatives, which is a subclass of \link[=new_pfselectstrat]{pfselectstrat}.
This is a uniform buy and hold strategy if not subclassing
(e.g. \link{new_best_stock})
}
\seealso{
new_buyandhold

Other buyandhold: 
\code{\link{best_stock}()},
\code{\link{new_buyandhold}()}
}
\concept{buyandhold}
