% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{check_strategy_args}
\alias{check_strategy_args}
\title{checks input args to a strategy}
\usage{
check_strategy_args(price_relatives, transaction_rate)
}
\arguments{
\item{price_relatives}{the price relatives to check}

\item{transaction_rate}{the transaction_rate to check}
}
\value{
No valid return
}
\description{
asserts that price relatives is a matrix with non-negative entries,
and if a column hits zero it stays zero.
asserts that transaction rate is a scalar double in \eqn{[0,1]}
}
