% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{combine_predictors}
\alias{combine_predictors}
\title{Combine Predictors for Lasso Model}
\usage{
combine_predictors(
  unpenalized_covariates,
  genetic_variants,
  interaction_terms,
  genetic_variants_knockoff,
  interaction_terms_knockoff
)
}
\arguments{
\item{unpenalized_covariates}{Matrix of covariates (not penalized).}

\item{genetic_variants}{Original genetic variants.}

\item{interaction_terms}{Interaction terms for genetic variants.}

\item{genetic_variants_knockoff}{Knockoff genetic variants.}

\item{interaction_terms_knockoff}{Interaction terms for knockoff genetic variants.}
}
\value{
A combined predictor matrix for model fitting.
}
\description{
Combine Predictors for Lasso Model
}
