% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{compute_price_relatives_whitener}
\alias{compute_price_relatives_whitener}
\title{Get whitener for each time step}
\usage{
compute_price_relatives_whitener(price_relative_matrix)
}
\arguments{
\item{price_relative_matrix}{A matrix of price relatives
with at least three rows.}
}
\value{
A list of whitening matrices, see description for details
}
\description{
Returns a list of matrices.
For \eqn{i >= 3}, the \eqn{i}th values and vectors are the
negative square root of the covariance matrix where we
treat the first \eqn{i-1} rows as observations and the columns as variables.
}
\details{
For \eqn{i=1,2} it's the identity matrix.
}
