% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{compute_prices_from_relatives}
\alias{compute_prices_from_relatives}
\alias{compute_price_matrix_from_relatives}
\title{get prices from price relatives}
\usage{
compute_prices_from_relatives(
  price_relatives,
  initial_price,
  .check_input = TRUE
)

compute_price_matrix_from_relatives(
  price_relative_matrix,
  initial_prices,
  .check_input = TRUE
)
}
\arguments{
\item{price_relatives}{a vector of price relatives
where each entry represents the price relative for
a trading period, i.e. price relative \eqn{x_t = p_{t+1} / p_t}.}

\item{initial_price}{\eqn{p_1}, the initial price.}

\item{.check_input}{If \code{TRUE}, checks that input meets the
given assumptions.}

\item{price_relative_matrix}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{p_{t+1} / p_t}, i.e.
the ratio of trading price to next price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{initial_prices}{A vector holding the initial prices of
each asset \eqn{i}}

\item{transaction_rate}{The percentage of each transaction (buy and sell)
spent on broker fees}
}
\value{
A vector with one more entry than \code{price_relatives}
     representing the true prices of the assets,
     the \eqn{i}th entry being the price during trading in period
     \eqn{i}, immediately before the price changes represent by price
     relatives \eqn{i} take place.
}
\description{
Given price_relatives and the initial prices, get the
actual prices
}
\section{Functions}{
\itemize{
\item \code{compute_price_matrix_from_relatives}: Computes a matrix of prices from a matrix
of price relatives, where each column represents an asset.
Output has same column names as \code{price_relative_matrix}
}}

