% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_metrics.R
\name{daily_return}
\alias{daily_return}
\alias{cumulative_wealth}
\title{Computes the daily increase in wealth}
\usage{
daily_return(price_relatives, portfolios_before_trade, transaction_rate)

cumulative_wealth(price_relatives, portfolios_before_trade, transaction_rate)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives}

\item{portfolios_before_trade}{a (T+1) x n matrix of the portfolios
before the trade (an entry in a portfolio is the portion
of total wealth in that asset)}

\item{transaction_rate}{The proportion of each buy/sell
lost to transaction fees, must be in \eqn{[0,1]}}
}
\value{
a vector of length T reporting the factor of wealth
     increase at each period
}
\description{
Given price relatives and portfolios at each period,
returns a vector containing the factor by
which cumulative wealth has increased at that trading period
}
\section{Functions}{
\itemize{
\item \code{cumulative_wealth}: Gives the factor by which initial wealth has increased at each trading
period.
}}

