% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_metrics.R
\name{evaluate_daily_return}
\alias{evaluate_daily_return}
\alias{evaluate_cumulative_wealth}
\title{Computes the daily increase in wealth}
\usage{
evaluate_daily_return(
  price_relatives,
  portfolios_after_trade,
  transaction_rate
)

evaluate_cumulative_wealth(
  price_relatives,
  portfolios_before_trade,
  transaction_rate
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives}

\item{portfolios_after_trade}{a T x n matrix of the portfolios
after the trade (an entry in a portfolio is the portion
of total wealth in that asset), the \eqn{i}th row
is after a trade in period \eqn{i} but before the \eqn{i}th
price relatives occur.}

\item{transaction_rate}{The proportion of each buy/sell
lost to transaction fees, must be in \eqn{[0,1]}}
}
\value{
a vector of length T whose
     \eqn{i}th entry reports the factor of wealth
     after the prices change according to the \eqn{i}th price relatives
}
\description{
Given price relatives and portfolios at each period,
returns a vector containing the factor by
which cumulative wealth has increased at that trading period
}
\section{Functions}{
\itemize{
\item \code{evaluate_cumulative_wealth}: Gives the factor by which initial wealth has increased at each trading
period.
}}

