% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{evaluate_momentum_data}
\alias{evaluate_momentum_data}
\title{Get momentum data from price relatives}
\usage{
evaluate_momentum_data(
  prices,
  price_means,
  time_window,
  consider_negative_momentum = TRUE
)
}
\arguments{
\item{consider_negative_momentum}{If \code{FALSE}, then
does not consider -1 as an option for momentum, i.e.
only checks historic mean and max of previous prices}

\item{price_relatives}{A numeric vector of the price}

\item{decay_factor}{The decay factor for the historic means
(as in \link{compute_historical_price_means})}
}
\value{
A tibble holding momentum and the trading period,
    as in the description
}
\description{
Given price relatives for a price and an alpha to use for
the historic price means, return a tibble with columns
}
