% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{evaluate_signed_momentum_at_window}
\alias{evaluate_signed_momentum_at_window}
\title{Return the signed momentum}
\usage{
evaluate_signed_momentum_at_window(price, next_price, historic_price_mean)
}
\arguments{
\item{price}{the current price (a scalar double)}

\item{next_price}{the next price (a scalar double)}

\item{historic_price_mean}{The current historic price mean (a scalar double)}
}
\value{
\code{0L} or \code{1L}, or \code{NA} if any of the arguments
    is missing
}
\description{
Returns \code{1L} if \code{next_price} is further from
\code{historic_price_mean} than \code{price}, and \code{0L} otherwise.
}
