% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{evaluate_unsigned_momentum}
\alias{evaluate_unsigned_momentum}
\title{Evaluates momentum on aggregated windows}
\usage{
evaluate_unsigned_momentum(agg_windows, momentum_window_size)
}
\arguments{
\item{agg_windows}{The \code{@eval{AGG_WINDOW_CLASS_NAME}} object
as returned from \code{\link{aggregate_price_and_mean_windows}}}

\item{momentum_window_size}{The size of the window used to determine
momentum. Must be at least 2 and no longer than
the price windows in \code{agg_windows}. Defaults to the
length of the price windows in \code{agg_windows}.}
}
\value{
An integer array whose \eqn{i}th entry holds the momentum
    corresponding to the \eqn{i}th row of \code{agg_windows}.
}
\description{
Given a \code{@eval{AGG_WINDOW_CLASS_NAME}} object as returned
from \code{\link{aggregate_price_and_mean_windows}}, computes
the momentum at each window (as in \code{evaluate_momentum_at_window})
}
