% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{exponential_gradient}
\alias{exponential_gradient}
\title{Create a \code{exponential_gradient} class}
\usage{
exponential_gradient(price_relatives, transaction_rate, learning_rate)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}
}
\value{
a \code{exponential_gradient} class
}
\description{
Returns an instance of a \code{exponential_gradient} class,
which is a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}.
See this paper
\url{http://rob.schapire.net/papers/HelmboldScSiWa98.pdf}
("Online Portfolio Selection Using Multiplicative Updates"
by David P Helmholt, Robert E. Schapire and Yoram Singer,
and Manfred K. Warmuth)
}
\seealso{
Other exponential_gradient: 
\code{\link{new_exponential_gradient}()},
\code{\link{validate_exponential_gradient}()}
}
\concept{exponential_gradient}
