% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{first_portfolio}
\alias{first_portfolio}
\title{Get first portfolio from strategy}
\usage{
first_portfolio(strategy, trading_period)
}
\arguments{
\item{strategy}{the strategy to use when determining the first portfolio.
Should be of class \link[=new_pfselectstrat]{pfselectstrat}}

\item{trading_period}{the "first" trading period, i.e. we might
    be starting backtesting in the middle of our data.

 @return A portfolio: a numeric vector with length equal to
     \code{strategy$nassets}, each entry representing the
     proportion of total wealth in the corresponding asset.}
}
\description{
Returns the first portfolio for the given strategy.
We assume trading is about to begin at the given trading period
The default asserts that the object is of class
\link[=new_pfselectstrat]{pfselectstrat}.
The default for class \link[=new_pfselectstrat]{pfselectstrat} is
a unifrom distribution of wealth across all assets.
}
\seealso{
Other pfselectstrat: 
\code{\link{new_pfselectstrat}()},
\code{\link{next_portfolio}()},
\code{\link{validate_pfselectstrat}()}
}
\concept{pfselectstrat}
