% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{fit_lasso_model}
\alias{fit_lasso_model}
\title{Fit Lasso Model with Cross-Validation}
\usage{
fit_lasso_model(X_matrix, y_vector, penalty_factors, n_folds = 5)
}
\arguments{
\item{X_matrix}{Matrix of predictors.}

\item{y_vector}{Response vector.}

\item{penalty_factors}{Vector of penalty factors.}

\item{n_folds}{Number of fold cross validation.}
}
\value{
Cross-validated Lasso model.
}
\description{
Fit Lasso Model with Cross-Validation
}
