% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{get_all_previous_price_windows}
\alias{get_all_previous_price_windows}
\title{Get all price windows of size window_size}
\usage{
get_all_previous_price_windows(prices, window_size)
}
\arguments{
\item{window_size}{the size of a previous price window.
Must be a postivie whole number}
}
\value{
A data frame with columns "asset" holding
    the column index of the price,
    "trading_period" holding the row index of the price,
    "price" holding the price during the trading period,
    and "previous_price_window": holding a list which
    contains either previous price window in order
    (least recent -> most recent).
    Rows will be be ordered by trading period, but other than
    that the order is undefined.
}
\description{
A previous price window of size w is the
w prices before the current price.
}
\details{
This function
returns a dtibble frame holding all the price windows of
size w for all assets in \code{prices}
}
