% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_with_Knockoff.R
\name{get_importance_matrices}
\alias{get_importance_matrices}
\title{Get Importance Matrices}
\usage{
get_importance_matrices(
  genetic_variants,
  genetic_variants_knockoff,
  additional_covariates,
  Z,
  y,
  n_folds = 5,
  FDR_rate = 0.1
)
}
\arguments{
\item{genetic_variants}{Matrix of genetic variants.}

\item{genetic_variants_knockoff}{Matrix of knockoff genetic variants.}

\item{Z}{Vector or matrix representing heterogeneity variable(s) (e.g., EUR or PCs).}

\item{y}{Outcome.}

\item{n_folds}{Number of fold cross validation with default of 5.}

\item{FDR_rate}{False Discovery Rate threshold, default is 0.1.}

\item{unpenalized_covariates}{Matrix of covariates (not penalized).}
}
\value{
A list containing \code{scaled_selection_matrix}, \code{selection_matrix}, and \code{W_statistic_matrix}.
}
\description{
This function generates scaled selection, selection, and W statistic matrices
using genetic variants and knockoff variants data.
}
