% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scrape.R
\name{get_investing_data}
\alias{get_investing_data}
\title{Get investing data at the url from the given dates}
\usage{
get_investing_data(url, st_date, end_date, message = FALSE)
}
\arguments{
\item{url}{The url to the historical data at investing.com}

\item{st_date}{the start date as a character in MM/DD/YYYY format}

\item{end_date}{the end date as a character in MM/DD/YYYY format}

\item{message}{If \code{TRUE}, messages the user with the results of the http query for the data. Defaults to \code{FALSE}.}
}
\description{
Given a url for historical data of a stock at investing.com,
extract a tibble with the stock data from \code{start_date}
to \code{end_date}
}
\examples{
"New Zealand" \%>\%
  msci_link() \%>\%
  get_investing_data('12/31/2019', '03/08/2020')
"https://www.investing.com/equities/american-airlines-group-historical-data" \%>\%
   get_investing_data('12/31/2019', '03/08/2020')
 "https://www.investing.com/equities/bank-of-america-historical-data" \%>\%
   get_investing_data('12/31/2019', '03/08/2020')

}
