% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{get_return_from_trade}
\alias{get_return_from_trade}
\title{Return the factor by which wealth increases}
\usage{
get_return_from_trade(
  tp_price_relatives,
  transaction_rate,
  prev_portfolio,
  portfolio,
  tol = 1e-10,
  maxit = 15,
  .check_input = TRUE
)
}
\arguments{
\item{tp_price_relatives}{A numeric vector whose \eqn{i}th entry
is the price relative for asset \eqn{i}, i.e. its price
at the new trading period divided by its price at the
previous trading period}

\item{transaction_rate}{The transaction rate: a scalar in \eqn{[0,1]}
which represents
the percentage of any stock transaction (buying and selling) which
goes towards transaction costs.}

\item{prev_portfolio}{The portfolio during the previous trading period
before the trade}

\item{portfolio}{The portfolio during the previous trading period
after the trade}

\item{tol}{We guarantee the trade is chosen so that
    the portfolio matches the argument \code{portfolio}
    within this tolerance (we have to iterate to
    reach a solution, since the trade is defined implicitly,
    see the comments inside the function)
@param maxit Stops after \code{maxit} iterations in
    used in computing a trade
@inheritParams compute_prices_from_relatives}
}
\value{
A numeric scalar representing the factor by which wealth
     increases, i.e. if \eqn{S_t} is wealth
     at trading period \eqn{t}, returns \eqn{\frac{S_t}{S_{t-1}}},
}
\description{
For given price relatives from trading period \eqn{(t-1)}->\eqn{t},
two portfolios (before and after a trade) at times \eqn{t-1}
and a transaction rate, return the factor by which total
wealth increases from time \eqn{t-1} to time \eqn{t}
}
\note{
both portfolios are
     numeric vectors whose \eqn{i}th entry represents
     the proportion of total wealth in asset \eqn{i}
     during the trading period \eqn{t-1}
}
