% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{historical_price_means}
\alias{historical_price_means}
\title{Return historical price means computed with decay factor}
\usage{
historical_price_means(prices, decay_factor = 0.5)
}
\description{
We define the historical price mean MA_t at time t
to be
\deqn{MA_t = decay_factor \cdot p_t + (1-decay_factor) \cdot MA_{t-1}}
And \eqn{MA_1 = p_1}.
}
\details{
@param prices A matrix of prices, each row is a time period
     and each column an asset
 @param decay_factor the decay factor for the mean (see description)
 @return A matrix with the same dimension as \code{prices}
     with historical price means as the entries

 @importFrom magrittr %>%
}
