% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric_matrix}
\alias{is_numeric_matrix}
\title{checks if x is a numeric matrix}
\usage{
is_numeric_matrix(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
TRUE iff x is a numeric matrix
}
\description{
checks if x is a numeric matrix
}
