% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_numeric_vector}
\alias{is_numeric_vector}
\title{checks if x is a numeric vector}
\usage{
is_numeric_vector(x)
}
\arguments{
\item{x}{the object to check}
}
\value{
TRUE iff x is a numeric vector
}
\description{
checks if x is a numeric vector
}
