% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_whole_number}
\alias{is_whole_number}
\title{tests whether x is close to an integer}
\usage{
is_whole_number(x)
}
\arguments{
\item{x}{the numeric scalar vector to test}
}
\value{
TRUE if x is close to an integer and
     false elsewhere
}
\description{
Test whether x is close to an integer
}
