% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{knockoff_filter}
\alias{knockoff_filter}
\title{Knockoff Filter Calculations}
\usage{
knockoff_filter(
  local_feature_importance,
  local_feature_importance_knockoff,
  FDR_rate = 0.1
)
}
\arguments{
\item{local_feature_importance}{Original feature importance matrix.}

\item{local_feature_importance_knockoff}{Knockoff feature importance matrix.}

\item{FDR_rate}{False Discovery Rate threshold.}
}
\value{
List containing the selection matrix (S_ij), scaled selection matrix, and variable info.
}
\description{
Knockoff Filter Calculations
}
