% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{momentum_confusion_matrix}
\alias{momentum_confusion_matrix}
\title{Compute a confusion matrix for price momentum}
\usage{
momentum_confusion_matrix(
  prices,
  historic_price_means,
  time_window,
  predicted_momentum,
  consider_negative_momentum = TRUE
)
}
\arguments{
\item{prices}{The prices during the trading periods (a matrix
with one more row than \code{price_relatives}).}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{predicted_momentum}{A matrix the same dimension as prices
holding the predicted momentum or NA}

\item{consider_negative_momentum}{If \code{FALSE}, then only
does not consider -1 as an option for momentum, i.e.
only checks historic mean and max of previous prices}
}
\value{
A confusion matrix for the momentum
}
\description{
Given prices and historical price means, as well as predicted momenta,
produce a confusion matrix for the momentum
}
