% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-scrape.R
\name{msci_link}
\alias{msci_link}
\title{Link to investing.com's historical data for a given MSCI country}
\usage{
msci_link(country_name)
}
\arguments{
\item{country_name}{The country name, e.g. \code{"New Zealand"}}
}
\value{
A string with the appropriate url
}
\description{
Given a country name, returns a string which we expect to
be the url at investing.com for the historical data
of that msci index.
}
\examples{
msci_link("Australia")
msci_link("New Zealand")
msci_link("Austria")

}
