% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nested_cv_regularized_pocket}
\alias{nested_cv_regularized_pocket}
\title{Uses nested cross-validation to pick parameters to \code{\link{regularized_pocket}}}
\usage{
nested_cv_regularized_pocket(
  x,
  y,
  weight_elimination,
  weight_decay,
  nfolds,
  maxit_per_fold
)
}
\arguments{
\item{x}{A numeric matrix with \eqn{n} rows. Should NOT include
a column of all 1s for bias weight.}

\item{y}{a numeric vector with \eqn{n} columns}

\item{weight_elimination}{a vector of weight elimination constants to try}

\item{weight_decay}{a vector of weight decay constants to try}

\item{nfolds}{Number of folds, if missing defaults to folds of size 1}

\item{maxit_per_fold}{maxits per fold}
}
\value{
a data frame with each combination of weight elimination
    and weight decay, as well as the recorded cv error
}
\description{
Tests all combinations of the weight elimination and weight decay
variables using nested cross validation (i.e. the nth row/entry of x/y
depends on the (n-1)th) and returns each pair with its associated
error
}
