% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_exponential_gradient}
\alias{new_exponential_gradient}
\title{Makes a \code{exponential_gradient} class}
\usage{
new_exponential_gradient(
  price_relatives,
  transaction_rate,
  learning_rate,
  ...,
  class = character()
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given \code{price_relatives}
     \code{nassets} and \code{ntrading_periods}}

\item{\eqn{\eta}}{in the referenced paper, the learning rate
should be a positive scalar, the higher the learning rate
the more volatile the algorithm}
}
\value{
a \code{exponential_gradient} class
}
\description{
Returns an instance of a \code{exponential_gradient} class,
which is a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}.
See this paper
\url{http://rob.schapire.net/papers/HelmboldScSiWa98.pdf}
("Online Portfolio Selection Using Multiplicative Updates"
by David P Helmholt, Robert E. Schapire and Yoram Singer,
and Manfred K. Warmuth)
}
\details{
has a multiplicative weight update rule, and initializes to
uniform portfolio

Validate using \code{\link{validate_expontial_gradient}()}.
}
\note{
As a user, instantiate using the \code{\link{exponential_gradient}()}
    function.
}
\seealso{
Other exponential_gradient: 
\code{\link{exponential_gradient}()},
\code{\link{validate_exponential_gradient}()}
}
\concept{exponential_gradient}
\keyword{internal}
