% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_loadonline}
\alias{new_loadonline}
\title{Implements the LOAD Online PS System}
\usage{
new_loadonline(
  price_relatives,
  transaction_rate,
  decay_factor,
  regularization_factor,
  time_window,
  momentum_threshold,
  wealth_factor_threshold,
  prices,
  price_means,
  ...,
  class = character()
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{decay_factor}{\eqn{\alpha} in the referenced paper, LOAD predicts
that stocks regressing to the mean have approximate price
\eqn{MA_t = \alpha p_t + (1-\alpha)MA_{t-1}}, \eqn{MA_1 = p_1}}

\item{regularization_factor}{\eqn{\lambda} in the referenced paper,
the regularization coeffecient for weight decay when
regressing the prices in the recent time window.
Must be non-negative}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{momentum_threshold}{\eqn{\eta} in the referenced paper.
If the regressed slope is greater than \eqn{\eta} we say the
stock has momentum. This must be greater than 0.}

\item{wealth_factor_threshold}{\eqn{\epsilon} in the referenced
paper. Once a price relative for the next
trading period is predicted, this is the minimum return
our traded portfolio should satisfy. This must be greater
than 0.}

\item{prices}{The prices during the trading periods (a matrix
with one more row than \code{price_relatives})}

\item{price_means}{The mean price \eqn{MA_t} is \code{decay_factor} * \eqn{p_t}
+ \code{(1-decay_factor)} * \eqn{MA_{t-1}}, a matrix the
same size as \code{prices}}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given \code{price_relatives}
     \code{nassets} and \code{ntrading_periods}}
}
\value{
A loadonline class object with the given values
}
\description{
A class to run the LOAD online PS System: For each stock
regresses the prices in the last time_window many trading
periods and thresholds the slope to classify the stock as
with momentum or without momentum. It uses this classification
to predict the price vector, then chooses the closest portfolio
which is predicted to return at least wealth_factor_threshold
over the next period
}
\details{
This method was developed in the following paper:
\url{https://www.sciencedirect.com/science/article/abs/pii/S0950705119303922#b19}

Returns a \code{loadonline} class object,
a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}

Initial portfolio is uniform.

For validation see \code{\link{validate_loadonline}()}.
}
\note{
Users should call the \code{\link{loadonline}} function instead.
}
\seealso{
Other loadonline: 
\code{\link{loadonline}()},
\code{\link{validate_loadonline}()}
}
\concept{loadonline}
\keyword{internal}
