% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_newton_step}
\alias{new_newton_step}
\title{Makes a \code{newton_step} class}
\usage{
new_newton_step(price_relatives, transaction_rate, ..., class = character())
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given \code{price_relatives}
     \code{nassets} and \code{ntrading_periods}}
}
\value{
a \code{newton_step} class
}
\description{
Returns an instance of a \code{newton_step} class,
which is a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}.
}
\details{
This is implements the Online Newton Step
(\url{https://www.researchgate.net/publication/221346006_Algorithms_for_portfolio_management_based_on_the_Newton_method})
from Agarwal, Hazan, Kale, & Schapire's 2006 paper
"Algorithms for portfolio management based on the Newton method".
Each step solves a small linear set of equation then projects to
the unit simplex in a matrix norm (estimated using projected
gradient descent).

The lhs and rhs of linear eqs build on each other, so have extra private
variables.
In the referenced paper, there are also parameters
\eqn{\eta}, \eqn{\beta}, and \eqn{\delta}. We set \eqn{\delta=1/8},
\eqn{\eta=0}, and \eqn{\beta=1}

initializes to uniform portfolio

Validate using \code{\link{validate_newton_step}()}.
}
\note{
As a user, instantiate using the \code{\link{newton_step}()}
    function.

has private extra slots .bt and .At which hold the
    \eqn{b_t} and \eqn{A_t} referenced in the paper
    for the most recent time
    perioc. These should NOT BE TOUCHED by the user,
    and values are \code{NA_real_} until computed.
}
\seealso{
Other newton_step: 
\code{\link{newton_step}()},
\code{\link{validate_newton_step}()}
}
\concept{newton_step}
\keyword{internal}
