% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_pfselectstrat}
\alias{new_pfselectstrat}
\title{makes a new \code{"pfselectstrat"} class}
\usage{
new_pfselectstrat(nassets, ntime_periods, ..., class = character())
}
\arguments{
\item{nassets}{A positive whole number, the number of assets}

\item{ntime_periods}{A positive whole number, the number of time periods.
The prices before the first trading period we think of as p_0,
and the first trading period occurs right before the prices change
to p_1}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given
     \code{nassets} and \code{ntime_periods}}
}
\description{
Creates a \code{"pfselectstrat"} class, which has a number
of assets and a number of time periods. We think of each
time period as corresponding to a change of prices. At the
beginning of the time period the prices have not yet changed,
and at the end of the time period they have changed to the next prices.
}
\details{
For validation of \code{pfselectstrat} objects, see
\code{\link{validate_pfselectstrat}()}
}
\note{
This is a class that should be built on, not instantiated
    and used directly by end users
}
\seealso{
Other pfselectstrat: 
\code{\link{first_portfolio}()},
\code{\link{next_portfolio}()},
\code{\link{validate_pfselectstrat}()}
}
\concept{pfselectstrat}
\keyword{internal}
