% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{newton_step}
\alias{newton_step}
\title{Create a \code{newton_step} class}
\usage{
newton_step(price_relatives, transaction_rate, variability_parameter)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{variability_parameter}{(OPTIONAL default 0)
 \eqn{\alpha} in the referenced paper,
the believed minimum price relative for any stock.}
}
\value{
a \code{newton_step} class
}
\description{
Returns an instance of a \code{newton_step} class,
which is a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}.
}
\seealso{
Other newton_step: 
\code{\link{new_newton_step}()},
\code{\link{validate_newton_step}()}
}
\concept{newton_step}
