% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{next_portfolio}
\alias{next_portfolio}
\title{Get next portfolio from strategy}
\usage{
next_portfolio(strategy, trading_period, portfolio)
}
\arguments{
\item{strategy}{the strategy to use when determining the first portfolio.
Should be of class \link[=new_pfselectstrat]{pfselectstrat}}

\item{trading_period}{the "first" trading period, i.e. we might
be starting backtesting in the middle of our data.}

\item{portfolio}{The portfolio at the beginning of this trading
    period (w.r.t. to the prices before the
    prices change according to the price relatives of this trading
    period)

 @return A portfolio: a numeric vector with length equal to
     \code{strategy$nassets}, each entry representing the
     proportion of total wealth in the corresponding asset.
     This should be the portfolio w.r.t. the current prices
     (before the prices change according to the price relatives
     of this trading periot)}
}
\description{
Return the next portfolio for the given strategy.
We assume trading is about to begin at the given trading period.
}
\seealso{
Other pfselectstrat: 
\code{\link{first_portfolio}()},
\code{\link{new_pfselectstrat}()},
\code{\link{validate_pfselectstrat}()}
}
\concept{pfselectstrat}
