% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyse_o.R
\docType{data}
\name{nyse_o}
\alias{nyse_o}
\title{NYSE(O)}
\format{
A matrix with 5651 rows and 36 columns of daily
   closing prices.
   Each row represents a trading day and each column
   represents an asset.
   Each entry is a price relative: if a stock has price
   \eqn{p_t} at trading period \eqn{t}, it has price relative
   \eqn{\frac{p_t}{p_{t-1}}} at trading period \eqn{t}.
   The column names are the stock names.
}
\source{
\url{http://www.cs.technion.ac.il/~rani/portfolios/portfolios.htm}
}
\usage{
nyse_o
}
\description{
New York Stock Exchange prices from 1962-1984.
Downloaded by Ben Sepanski
}
\details{
\describe{
  \item{Market:}{New York Stock Exchange}
  \item{Number of stocks}{36}
  \item{Number of days}{5651}
  \item{Number of years}{22}
  \item{Period}{Jul 3, 1962 to Dec 31, 1984}
  \item{Other Info}{Contains very large market cap stocks.
                    Dataset was collected by Hal Stern}
}
}
\keyword{datasets}
