% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{onlineLOAD}
\alias{onlineLOAD}
\title{creates a onlineLOAD strategy}
\usage{
onlineLOAD(
  price_relatives,
  transaction_rate,
  decay_factor,
  regularization_factor,
  time_window,
  momentum_threshold,
  wealth_factor_threshold,
  prices,
  price_means,
  check_prices = TRUE,
  check_price_means = TRUE
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{decay_factor}{\eqn{\alpha} in the referenced paper, LOAD predicts
that stocks regressing to the mean have approximate price
\eqn{MA_t = \alpha p_t + (1-\alpha)MA_{t-1}}, \eqn{MA_1 = p_1}}

\item{regularization_factor}{\eqn{\lambda} in the referenced paper,
the regularization coeffecient for weight decay when
regressing the prices in the recent time window.
Must be non-negative}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{momentum_threshold}{\eqn{\eta} in the referenced paper.
If the regressed slope is greater than \eqn{\eta} we say the
stock has momentum. This must be greater than 0.}

\item{wealth_factor_threshold}{\eqn{\epsilon} in the referenced
paper. Once a price relative for the next
trading period is predicted, this is the minimum return
our traded portfolio should satisfy. This must be greater
than 0.}

\item{prices}{(OPTIONAL), initial_prices default to 1 if missing.
See \code{\link{new_onlineLOAD}} for a description of how to pass them in.
Otherwise, they are computed on construction.}

\item{price_means}{(OPTIONAL)
See \code{\link{new_onlineLOAD}} for a description of how to pass them in.
Otherwise, they are computed on construction.}

\item{check_prices}{checks prices match price relatives iff this is
\code{TRUE}. This can be expensive.}

\item{check_price_means}{checks price means match prices iff this is
\code{TRUE}. This can be expensive.}
}
\value{
A onlineLOAD class object with the given values
}
\description{
A class to run the LOAD online PS System: For each stock
regresses the prices in the last time_window many trading
periods and thresholds the slope to classify the stock as
with momentum or without momentum. It uses this classification
to predict the price vector, then chooses the closest portfolio
which is predicted to return at least wealth_factor_threshold
over the next period
}
\seealso{
Other onlineLOAD: 
\code{\link{new_onlineLOAD}()},
\code{\link{validate_onlineLOAD}()}
}
\concept{onlineLOAD}
