% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lasso_with_Knockoff.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot Heatmap of Selected SNPs Including Knockoff}
\usage{
plot_heatmap(
  scaled_selection_matrix,
  genetic_variants,
  selected_snp_names,
  eur,
  save_path = "."
)
}
\arguments{
\item{scaled_selection_matrix}{Scaled selection matrix from the knockoff filter.}

\item{selected_snp_names}{Names of selected SNPs.}

\item{eur}{Vector representing EUR percentages to sort individuals.}

\item{save_path}{The directory path where the plot will be saved. Defaults to the current working directory.}
}
\value{
A ggplot object representing the heatmap.
}
\description{
Plot Heatmap of Selected SNPs Including Knockoff
}
