% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{plot_pcs}
\alias{plot_pcs}
\title{Plot PCs Visualization for a SNP}
\usage{
plot_pcs(
  pcs,
  snp_importance,
  snp_name,
  pc_x = "PC1",
  pc_y = "PC2",
  save_path = "."
)
}
\arguments{
\item{pcs}{Principal Components data (a matrix or data frame with at least two columns, such as PC1, PC2, etc.).}

\item{snp_importance}{Numeric vector representing feature importance values for the selected SNP.}

\item{snp_name}{Name of the SNP to visualize.}

\item{pc_x}{The name of the principal component to plot on the x-axis (e.g., "PC1").}

\item{pc_y}{The name of the principal component to plot on the y-axis (e.g., "PC2").}

\item{save_path}{The directory path where the plot will be saved. Defaults to the current working directory.}
}
\value{
A ggplot object of the specified PCs colored by feature importance.
}
\description{
Plot PCs Visualization for a SNP
}
