% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{predict_momentum_LOAD}
\alias{predict_momentum_LOAD}
\title{Test momentum prediction using LOAD strategy}
\usage{
predict_momentum_LOAD(
  aggregated_momentum_data,
  regularization_factor,
  momentum_threshold,
  max_cv_window = NULL,
  momentum_colname = "nonnegative_momentum"
)
}
\arguments{
\item{aggregated_momentum_data}{A tibble
of class \code{@eval{AGG_WINDOW_CLASS_NAME}} as returned from
\code{\link{aggregate_price_and_mean_windows}}, with an
added column of column name \code{momentum_colname}
holding the momentum.}

\item{regularization_factor}{A regularization factor used in the
LOAD prediction, if a vector then cross-validates
(at each step choosing the factor that had the
least error over previous time, random first time)}

\item{momentum_threshold}{The threshold to be classified as
with momentum. If a vector then cross-validates.
(at each step choosing the factor that had the
least error over previous time, random first time)}

\item{max_cv_window}{If \code{NULL} then checks over all
past error, otherwise only checks the past \code{max_cv_window}
time steps to determine error.}

\item{momentum_colname}{The name of the column holding the
true momentum values (as a string),
defaults to \code{"nonnegative_momentum"}.}
}
\value{
\code{aggregated_momentum_data}
    with added columns:
    \code{LOAD_prediction} holding LOAD's predicted
    momentum for that time, \code{regularization_factor}
    holding the regularization factor used in that prediction,
    \code{momentum_threshold} holding the
    momentum threshold used in that prediction.
}
\description{
Test LOAD's strategy of predicting momentum. Returns
each momentum prediction.
See \code{\link{evaluate_momentum_at_window}} for a description
of momentum
}
