% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{predict_momentum_pocket}
\alias{predict_momentum_pocket}
\title{Predicts momentum using a regularized pocket algorithm}
\usage{
predict_momentum_pocket(
  data,
  feature_colname,
  momentum_colname = "momentum",
  cv_assets,
  weight_elimination,
  maxit,
  max_cv_window = NULL
)
}
\arguments{
\item{data}{A tibble
of class \code{@eval{AGG_WINDOW_CLASS_NAME}} as returned from
\code{\link{aggregate_price_and_mean_windows}}, with an
added column of column name \code{momentum_colname}
holding the momentum.}

\item{feature_colname}{The column name of the features to use.
Should be a list column, each list entry holding a double vector.}

\item{momentum_colname}{The name of the column holding the
true momentum values (as a string),
defaults to \code{"momentum"}. Should be (possibly negative) momentum
values (see \code{\link{evaluate_momentum_at_window}})}

\item{weight_elimination}{A double vector, at each stage picks the
one which has performed best in the past cv window (or all of the
past, whichever is selected)}

\item{maxit}{the maximum number of iterations}

\item{max_cv_window}{If \code{NULL} then checks over all
past error, otherwise only checks the past \code{max_cv_window}
time steps to determine error.}
}
\value{
A tibble like \code{data} but with a \code{pocket_prediction}
    \code{pocket_error}, and \code{pocket_elimination_weight},
    columns holding the momentum prediction,
    whether it was an error, and the elimination weight used.
}
\description{
See \code{\link{regularized_pocket}()} for a description of the
algorithm used.
Either uses the \code{weight_elimination} constant given
or cross-validates at each trading period to choose one of
the selected.
If there is at least one cv asset,
then each stage runs once with previous weights and once from linear
regression weights, then picks whichever does best on cv error
}
