% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{predict_momentum_reg_pocket}
\alias{predict_momentum_reg_pocket}
\title{Predict the momentum using a regularized pocket algorithm}
\usage{
predict_momentum_reg_pocket(
  prices,
  time_window,
  price_means,
  weight_elimination,
  trading_period_probs,
  maxit,
  num_train
)
}
\arguments{
\item{price_means}{the historic mean of the prices (exponentially
weighted average of past prices)}

\item{weight_elimination}{the weight elimination factor to use}

\item{trading_period_probs}{The relative probability of each trading period,
used for sampling and error evaluation}

\item{maxit}{the max number of iterations when training}

\item{num_train}{the number of initial trading periods to train on}
}
\description{
Trains on the first
num_train parameters from the given weight elimination parameters.
Then, predicts the remaining momenta.
}
