% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{predict_price_LOAD}
\alias{predict_price_LOAD}
\title{predicts price according to LOAD strategy}
\usage{
predict_price_LOAD(
  prev_prices,
  historic_mean,
  regularization_factor,
  momentum_threshold,
  min_var = .Machine$double.eps^0.25
)
}
\arguments{
\item{prev_prices}{a numeric vector of previous prices}

\item{historic_mean}{The historic mean of the prices}

\item{regularization_factor}{\eqn{\lambda} in the referenced paper,
the regularization coeffecient for weight decay when
regressing the prices in the recent time window.
Must be non-negative}

\item{momentum_threshold}{\eqn{\eta} in the referenced paper.
If the regressed slope is greater than \eqn{\eta} we say the
stock has momentum. This must be greater than 0.}

\item{min_var}{(OPTIONAL) minimum variance to be determined non-constant}
}
\value{
the predicted next price
}
\description{
predicts next prices given previous prices according to LOAD
strategy (see paper mentioned in description of
\code{\link{backtest_LOAD}})
}
\note{
NO TYPE CHECKING IS PERFORMED... be careful
}
