% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{prepare_interaction_terms}
\alias{prepare_interaction_terms}
\title{Prepare Interaction Terms}
\usage{
prepare_interaction_terms(genetic_variants, Z)
}
\arguments{
\item{genetic_variants}{Matrix of original or knockoff genetic variants.}

\item{Z}{Vector or matrix representing heterogeneity variable(s) (e.g., EUR or 4 PCs).}
}
\value{
Interaction terms for genetic variants.
}
\description{
Prepare Interaction Terms
}
