% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{price_adjusted_portfolio}
\alias{price_adjusted_portfolio}
\title{renormalizes portfolio to show total portion of wealth in each asset}
\usage{
price_adjusted_portfolio(price_relatives, portfolio)
}
\arguments{
\item{price_relatives}{A vector whose \eqn{i}th entry
is the ratio of the \eqn{i}th asset's new price
to its old price. More precisely, if asset \eqn{i}
had price \eqn{p_{t-1}} when wealth was distributed
according to \code{portfolio}, but now has price \eqn{p_t},
the \eqn{i}th entry of \code{price_relatives} is
\eqn{\frac{p_t}{p_{t-1}}}.}

\item{portfolio}{A vector of non-negative numbers which sum to 1,
the \eqn{i}th entry representing the portion of total wealth
in asset \eqn{i}.}
}
\value{
A numeric vector whose \eqn{i}th entry is the portion of total
     wealth in asset \eqn{i} (according to the new prices).
}
\description{
Starting with a portfolio whose \eqn{i}th entry was the portion
of ones total wealth in asset \eqn{i} for the previous prices,
return a new vector reflecting the portion of total wealth
in each asset under the new prices.
}
