% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_to_simplex}
\alias{project_to_simplex}
\title{Computes Euclidean Projection onto simplex}
\usage{
project_to_simplex(v, r = 1)
}
\arguments{
\item{v}{The vector to project}

\item{r}{(DEFAULT 1) radius of simplex}
}
\value{
the projection of \code{v} onto the \code{r}-simplex
}
\description{
Returns the closest vector on the simplex of radius
r
}
\details{
Returns minimizer of
\deqn{\frac{1}{2}||w-v||^2,  v^T1 = r, v >= 0}
using method from the following paper:
\url{https://stanford.edu/~jduchi/projects/DuchiShSiCh08.pdf}.
}
