% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{project_to_simplex_A_norm}
\alias{project_to_simplex_A_norm}
\title{project to simplex in the A norm}
\usage{
project_to_simplex_A_norm(
  v,
  A,
  r = 1,
  step_sizes = c(1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001),
  maxit = 5,
  tol = 1e-07
)
}
\arguments{
\item{v}{The vector to project}

\item{A}{The positive-definite matrix to use for a norm
(NOTE: no checks performed to verify)}

\item{r}{(DEFAULT 1) radius of simplex}

\item{step_sizes}{(OPTIONAL) Projected gradient descent step size.
If a vector, tries each step size and picks closest.
Should be sorted greatest -> smallest (descending)}

\item{maxit}{(OPTIONAL) max number of iterations}

\item{tol}{(OPTIONAL) Min update before termination}
}
\description{
Find the closest vector on the simplex to v,
where closest is in terms of the A norm
\eqn{x^TAx}
Runs projected gradient descent for maxit
or until tolernace is reached
}
