% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdirichlet_onehalf}
\alias{rdirichlet_onehalf}
\title{Samples from Dirichlet(1/2,1/2,...,1/2)}
\usage{
rdirichlet_onehalf(n, d)
}
\arguments{
\item{n}{the number of samples}

\item{d}{the dimension of each sample}
}
\value{
An n x d matrix
     of \code{n} samples from the \code{d}-dimensional
     Dirichlet(1/2,1/2,...,1/2) distribution
     (each sample is a row)
}
\description{
Returns independent samples from the Dirichlet(1/2,1/2,...,1/2)
distribution
}
\seealso{
\url{https://www.wikiwand.com/en/Dirichlet_distribution#/Random_number_generation}
}
