% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{rebase_agg_windows}
\alias{rebase_agg_windows}
\title{Rebase to new linear combination of assets}
\usage{
rebase_agg_windows(
  agg_windows,
  new_assets,
  asset_rownames,
  scalar_columns_to_rebase = NULL
)
}
\arguments{
\item{agg_windows}{A tibble
of class \code{@eval{AGG_WINDOW_CLASS_NAME}} as returned from
\code{\link{aggregate_price_and_mean_windows}}}

\item{new_assets}{A list, whose \eqn{i}th entry is a square matrix.
Each column represents a new
linear combination of the assets to consider replacing
one of the old assets.}

\item{asset_rownames}{These are the names of the assets of
\code{agg_windows} in the order which they correspond to
the rows of the matrices in \code{new_assets}}

\item{scalar_columns_to_rebase}{A character vector of
column names of \code{agg_windows} to be rebased. These
must be double columns.}
}
\value{
the new tibble with rebased values
}
\description{
Change the price windows and historic price mean windows
in trading period \eqn{i} to represent a
linear combination of prices, the \eqn{j}th combination being the \eqn{j}th
column of the matrix which is the \eqn{i}th entry of \code{new_assets}.
}
