% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{rebase_price_relative_matrix}
\alias{rebase_price_relative_matrix}
\title{Rebase price relatives to new linear combination}
\usage{
rebase_price_relative_matrix(price_relative_matrix, new_assets)
}
\arguments{
\item{price_relative_matrix}{The matrix of price relatives}

\item{new_assets}{A list, whose \eqn{i}th entry is a square matrix of size
\code{ncol(price_relative_matrix)}. Each column represents a new
linear combination of the assets to consider.}
}
\value{
the new price relative matrix
}
\description{
Change the price relatives in trading period \eqn{i} to represent
the price relatives of a
linear combination of prices, the \eqn{j}th combination being the \eqn{j}th
column of the matrix which is the \eqn{i}th entry of \code{new_assets}.
}
