% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{rescale_to_price_over_mean}
\alias{rescale_to_price_over_mean}
\title{Rescales windows to positive slope}
\usage{
rescale_to_price_over_mean(agg_windows)
}
\arguments{
\item{agg_windows}{a \code{@eval{AGG_WINDOW_CLASS_NAME} } instance
as returned from \code{\link{aggregate_price_and_mean_windows}}.}
}
\value{
\code{agg_windows} but modified to have an extra column
    \code{flip_sign} which is -1 if the sign was flipped and 1 otherwise
    as well as having rescaled
    the \code{price_window} and \code{historic_price mean_window}
    columns by \code{flip_sign}
}
\description{
Rescales windows (\code{price_window} and \code{historic_price_mean_window})
by \eqn{\pm 1} so that each the most recent price is greater than
or equal to the most recen thistoric price mean.
}
