% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rollify}
\alias{rollify}
\alias{rollify_dbl}
\title{rollify f to operate over windows}
\usage{
rollify(f, window_sizes, fill = NA)

rollify_dbl(f, window_sizes, fill = NA)
}
\arguments{
\item{f}{the function to roll.}

\item{window_sizes}{the size of the windows in each parameter.
Must be at least one.}

\item{fill}{the value to fill in entries when there isn't
enough data for a given window size}
}
\value{
a function which rolls \code{f} over windows of the
    given sizes.
}
\description{
Returns a function which rolls \code{f} over windows.
\code{window_sizes} is a vector,
\code{window_sizes[i]} should be the window size for the \eqn{i}th
positional argument of \code{f}. This function can only
roll positional arguments.
}
\details{
For example, if we have f(x,y,z) where x, y, and z are some windows,
and we want to roll f across vectors (xvec, yvec, zvec), we
would call rollify(f)(xvec, yvec, zvec). These vecs
must all have the same length, and there must be at least one.
}
\section{Functions}{
\itemize{
\item \code{rollify_dbl}: \code{rollify} is to \code{rollify_dbl} as \code{purrr::map}
is to \code{purrr::map_dbl}. i.e. for \code{rollify_dbl}, \code{f}
\emph{MUST} return a single numeric value.
Instead of returning a list, the function returned
by \code{rollify_dbl} returns a numeric vector.
}}

