% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{set_penalty_factors}
\alias{set_penalty_factors}
\title{Set Penalty Factors}
\usage{
set_penalty_factors(
  unpenalized_covariates,
  genetic_variants,
  interaction_terms,
  genetic_variants_knockoff,
  interaction_terms_knockoff
)
}
\arguments{
\item{unpenalized_covariates}{Matrix of covariates.}

\item{genetic_variants}{Matrix of original genetic variants.}

\item{interaction_terms}{Interaction terms for genetic variants.}

\item{genetic_variants_knockoff}{Knockoff genetic variants.}

\item{interaction_terms_knockoff}{Interaction terms for knockoff genetic variants.}
}
\value{
Vector of penalty factors.
}
\description{
Set Penalty Factors
}
