% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{uniform_portfolio}
\alias{uniform_portfolio}
\title{Returns a uniform portfolio}
\usage{
uniform_portfolio(nassets)
}
\arguments{
\item{nassets}{the number of assets}
}
\value{
a numeric vector of length \code{nassets}, each
    with value \code{1/nassets}.
}
\description{
Returns a uniform portfolio over the given number of assets
}
