% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{universal_portfolio}
\alias{universal_portfolio}
\title{Create a \code{universal_portfolio} class}
\usage{
universal_portfolio(
  price_relatives,
  transaction_rate,
  nsamples,
  consider_transaction_rate
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{nsamples}{Number of samples to perform at each day.
In general, this will be lower than theoretically necessary,
so expect a sub-optimal estimate}

\item{consider_transaction_rate}{If \code{FALSE}, uses the original
1991 algorithm proposed by Cover, where the cumulative wealth
produced by a constantly rebalanced portfolio (CRP) is
computed without consider transaction costs. If \code{TRUE},
follows Blum & Kalai's 1991 paper and considers the transaction
costs.}
}
\value{
a \code{universal_portfolio} class
}
\description{
Returns an instance of a \code{universal_portfolio} class,
which is a subclass of \code{\link[=new_pfselectstrat]{pfselectstrat}}.
See Cover's paper "Universal Portfolios" (1991,
\url{https://onlinelibrary.wiley.com/doi/epdf/10.1111/j.1467-9965.1991.tb00002.x}
) for a description.
We estimate the universal portfolio using
a sampling estimation method mentioned in Blum & Kalai's
1999 paper "Universal Portfolios With and Without Transaction Costs"
(\url{https://link.springer.com/article/10.1023/A:1007530728748}),
but is basically just random sampling of portfolios, then the
average weighted by the produced cumulative wealth up to this point.
However, we use Dirichlet(1/2,1/2,...,1/2) priors for the portfolios
because they may have better performance according to
Cover & Ordentlich (1996)
}
\seealso{
Other universal_portfolio: 
\code{\link{new_universal_portfolio}()},
\code{\link{validate_universal_portfolio}()}
}
\concept{universal_portfolio}
