% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_buyandhold}
\alias{validate_buyandhold}
\title{validates a buyandhold object}
\usage{
validate_buyandhold(x)
}
\arguments{
\item{x}{the \code{"buyandhold"} instance to validate}
}
\description{
Asserts that the dimensions of the price_relatives matches
the number of trading period and number of assets,
that all the price relatives are non-negative,
as well as type checking. Calls \code{\link{validate_pfselectstrat}}.
}
\seealso{
Other buyandhold: 
\code{\link{best_stock}()},
\code{\link{new_buyandhold}()},
\code{\link{uniform_bah}()}
}
\concept{buyandhold}
