% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_exponential_gradient}
\alias{validate_exponential_gradient}
\title{validates a \code{\link[=new_exponential_gradient]{exponential_gradient}}
class}
\usage{
validate_exponential_gradient(x)
}
\arguments{
\item{x}{the \code{\link[=new_exponential_gradient]{exponential_gradient}}
object to validate}
}
\value{
the object x
}
\description{
validates the class, making sure learning rate is positive
and following checks in \code{\link{validate_pfselectstrat}}
}
\seealso{
Other exponential_gradient: 
\code{\link{exponential_gradient}()},
\code{\link{new_exponential_gradient}()}
}
\concept{exponential_gradient}
