% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_loadonline}
\alias{validate_loadonline}
\title{validate a loadonline class}
\usage{
validate_loadonline(x)
}
\arguments{
\item{x}{the loadonline object to validate}
}
\value{
the object x
}
\description{
Given a loadonline class (see \code{\link{new_loadonline}})
validates that any numeric constants are scalars,
if they should be non-negative or in \eqn{[0,1]} then they are,
that the price means are correct, and the prices match the price
relatives
}
\seealso{
Other loadonline: 
\code{\link{loadonline}()},
\code{\link{new_loadonline}()}
}
\concept{loadonline}
