% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_newton_step}
\alias{validate_newton_step}
\title{validates a \code{\link[=new_newton_step]{newton_step}}
class}
\usage{
validate_newton_step(x)
}
\arguments{
\item{x}{the \code{\link[=new_newton_step]{newton_step}}
object to validate}
}
\value{
the object x
}
\description{
validates the class, making sure \code{variability_parameter}
is a non-negative number, and checks \code{.At} and \code{.bt}.
Follows checks in \code{\link{validate_pfselectstrat}}
}
\seealso{
Other newton_step: 
\code{\link{new_newton_step}()},
\code{\link{newton_step}()}
}
\concept{newton_step}
