% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_nonnegative_mat}
\alias{validate_nonnegative_mat}
\title{Validates and returns a numeric matrix}
\usage{
validate_nonnegative_mat(mat)
}
\arguments{
\item{mat}{the object to validate}
}
\value{
Returns \code{mat}
}
\description{
Asserts that \code{mat} is a numeric matrix
with non-negative entries.
Returns \code{portfolio}
}
