% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_onlineLOAD}
\alias{validate_onlineLOAD}
\title{validate a onlineLOAD class}
\usage{
validate_onlineLOAD(x, check_prices = TRUE, check_price_means = TRUE)
}
\arguments{
\item{x}{the onlineLOAD object to validate}

\item{check_prices}{checks prices match price relatives iff this is
\code{TRUE}. This can be expensive.}

\item{check_price_means}{checks price means match prices iff this is
\code{TRUE}. This can be expensive.}
}
\value{
the object x
}
\description{
Given a onlineLOAD class (see \code{\link{new_onlineLOAD}})
validates that any numeric constants are scalars,
if they should be non-negative or in \eqn{[0,1]} then they are,
that the price means are correct, and the prices match the price
relatives
}
\seealso{
Other onlineLOAD: 
\code{\link{new_onlineLOAD}()},
\code{\link{onlineLOAD}()}
}
\concept{onlineLOAD}
