% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_pfselectstrat}
\alias{validate_pfselectstrat}
\title{validates a \code{"pf_selectstrat"} class}
\usage{
validate_pfselectstrat(x)
}
\arguments{
\item{x}{the \code{"pf_selectstrat"} instance to validate}
}
\description{
Asserts that the given object has values \code{nassets}
and \code{ntime_periods} which are both positive whole numbers
}
\seealso{
Other pfselectstrat: 
\code{\link{first_portfolio}()},
\code{\link{new_pfselectstrat}()},
\code{\link{next_portfolio}()}
}
\concept{pfselectstrat}
