% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{validate_portfolio}
\alias{validate_portfolio}
\title{Validates and returns a portfolio}
\usage{
validate_portfolio(portfolio, nassets)
}
\arguments{
\item{portfolio}{the portfolio to validate}

\item{nassets}{the number of assets the portfolio should have}
}
\value{
portfolio
}
\description{
Asserts that \code{portfolio} is a numeric vector
of length \code{nassets} with non-negative entries
that sum to one. Returns \code{portfolio}
}
