% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{validate_portfolio_matrix}
\alias{validate_portfolio_matrix}
\title{validate and return a matrix of portfolios}
\usage{
validate_portfolio_matrix(portfolio_matrix, ntrading_periods, nassets)
}
\arguments{
\item{portfolio_matrix}{a matrix where each row is a portfolios}

\item{ntrading_periods}{the number of trading periods, \code{portfolio_matrix}
should have one more than this many rows}

\item{nassets}{the number of assets, should be the number of columns
in \code{portfolio_matrix}}
}
\value{
portfolio_matrix
}
\description{
validate and return a matrix of portfolios
}
