% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{validate_universal_portfolio}
\alias{validate_universal_portfolio}
\title{validates a \code{\link[=new_universal_portfolio]{universal_portfolio}}
class}
\usage{
validate_universal_portfolio(x)
}
\arguments{
\item{x}{the \code{\link[=new_universal_portfolio]{universal_portfolio}}
object to validate}
}
\value{
the object x
}
\description{
validates the class, making sure \code{nsamples} is a positive
whole number and that consider_transaction_rate is a single boolean.
Follows checks in \code{\link{validate_pfselectstrat}}
}
\seealso{
Other universal_portfolio: 
\code{\link{new_universal_portfolio}()},
\code{\link{universal_portfolio}()}
}
\concept{universal_portfolio}
