% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_exponential_gradient}
\alias{backtest_exponential_gradient}
\title{Backtests using a exponential gradient rule}
\usage{
backtest_exponential_gradient(
  price_relative_matrix,
  transaction_rate,
  learning_rate
)
}
\arguments{
\item{learning_rate}{The learning rate (\eqn{\eta} in the referenced
paper), a higher learning rate corresponds to a higher
sensitivitiy as one would expect.}
}
\description{
backtests using an exponential gradient rule.
See this paper
\url{http://rob.schapire.net/papers/HelmboldScSiWa98.pdf}
("Online Portfolio Selection Using Multiplicative Updates"
by David P Helmholt, Robert E. Schapire and Yoram Singer,
and Manfred K. Warmuth)
}
\details{
has a multiplicative weight update rule, and initializes to
uniform portfolio
}
