% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_LOAD}
\alias{backtest_LOAD}
\title{Backtests the LOAD Online PS System}
\usage{
backtest_LOAD(
  price_relatives,
  transaction_rate,
  decay_factor,
  regularization_factor,
  time_window,
  momentum_threshold,
  wealth_factor_threshold,
  prices,
  price_means
)
}
\arguments{
\item{price_relatives}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{\frac{p_t}{p_{t-1}}}, i.e.
the ratio of previous price to new price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{transaction_rate}{The percentage of each transaction (buy and sell)
spent on broker fees}

\item{decay_factor}{\eqn{\alpha} in the referenced paper, LOAD predicts
that stocks regressing to the mean have approximate price
\eqn{MA_t = \alpha p_t + (1-\alpha)MA_{t-1}}, \eqn{MA_1 = p_1}}

\item{regularization_factor}{\eqn{\lambda} in the referenced paper,
the regularization coeffecient for weight decay when
regressing the prices in the recent time window.
Must be non-negative}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{momentum_threshold}{\eqn{\eta} in the referenced paper.
If the regressed slope is greater than \eqn{\eta} we say the
stock has momentum. This must be greater than 0.}

\item{wealth_factor_threshold}{\eqn{\epsilon} in the referenced
paper. Once a price relative for the next
trading period is predicted, this is the minimum return
our traded portfolio should satisfy. This must be greater
than 0.}

\item{prices}{(OPTIONAL) The prices during the trading periods (a matrix
with one more row than \code{price_relatives}).
If not included, initial prices are assumed to be one.}

\item{price_means}{(OPTIONAL)
The mean price \eqn{MA_t} is \code{decay_factor} * \eqn{p_t}
+ \code{(1-decay_factor)} * \eqn{MA_{t-1}}, a matrix the
same size as \code{prices}. If not used, these are computed
from the prices}
}
\description{
Runs the LOAD online PS System: For each stock
regresses the prices in the last time_window many trading
periods and thresholds the slope to classify the stock as
with momentum or without momentum. It uses this classification
to predict the price vector, then chooses the closest portfolio
which is predicted to return at least wealth_factor_threshold
over the next period
}
\details{
This method was developed in the following paper:
\url{https://www.sciencedirect.com/science/article/abs/pii/S0950705119303922#b19}

Initial portfolio is uniform.
}
