% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_market}
\alias{backtest_market}
\title{backtests a market (uniform buy and hold) strategy}
\usage{
backtest_market(price_relatives, transaction_rate)
}
\arguments{
\item{price_relatives}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{\frac{p_t}{p_{t-1}}}, i.e.
the ratio of previous price to new price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{transaction_rate}{The percentage of each transaction (buy and sell)
spent on broker fees}
}
\description{
Backtests a strategy which buys an equal amount of all stocks
and then holds them.
}
