% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_online_newton_step}
\alias{backtest_online_newton_step}
\title{Backtests using an online newton step method}
\usage{
backtest_online_newton_step(price_relatives, transaction_rate)
}
\arguments{
\item{price_relatives}{a matrix of price relatives, each row
representing a trading period and each column an asset.
A price relative is \eqn{\frac{p_t}{p_{t-1}}}, i.e.
the ratio of previous price to new price. Prices change
according to the price relatives after the trade,
i.e. the price relatives for the trading period are not known
at trading time}

\item{transaction_rate}{The percentage of each transaction (buy and sell)
spent on broker fees}
}
\description{
Backtests using an online newton step method for portfolio
selection.
}
\details{
This is implements the Online Newton Step
(\url{https://www.researchgate.net/publication/221346006_Algorithms_for_portfolio_management_based_on_the_Newton_method})
from Agarwal, Hazan, Kale, & Schapire's 2006 paper
"Algorithms for portfolio management based on the Newton method".
Each step solves a small linear set of equation then projects to
the unit simplex in a matrix norm (estimated using projected
gradient descent).

According to the paper, we set \eqn{\eta=0}, \eqn{\beta=1}, and
\eqn{\delta=1/8}

initializes to uniform portfolio
}
