% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_strategy.R
\name{backtest_online_newton_step}
\alias{backtest_online_newton_step}
\title{Backtests using an online newton step method}
\usage{
backtest_online_newton_step(price_relative_matrix, transaction_rate)
}
\description{
Backtests using an online newton step method for portfolio
selection.
}
\details{
This is implements the Online Newton Step
(\url{https://www.researchgate.net/publication/221346006_Algorithms_for_portfolio_management_based_on_the_Newton_method})
from Agarwal, Hazan, Kale, & Schapire's 2006 paper
"Algorithms for portfolio management based on the Newton method".
Each step solves a small linear set of equation then projects to
the unit simplex in a matrix norm (estimated using projected
gradient descent).

According to the paper, we set \eqn{\eta=0}, \eqn{\beta=1}, and
\eqn{\delta=1/8}

initializes to uniform portfolio
}
