% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtesting.R
\name{backtest_portfolio_selector}
\alias{backtest_portfolio_selector}
\title{backtest to get cumulative wealth and portfolios over time}
\usage{
backtest_portfolio_selector(
  strategy,
  first_trading_period,
  last_trading_period,
  check_portfolio = TRUE
)
}
\arguments{
\item{strategy}{the \link[=new_pfselectstrat]{pfselectstrat} instance to use}

\item{first_trading_period}{the trading period to start backtesting
on (right before prices change)}

\item{last_trading_period}{(OPTIONAL) if supplied, only backtests
up through \code{max_trading_period} (instead of through all
available trading periods). This is the last price change we experience}

\item{check_portfolio}{(OPTIONAL) defaults to \code{TRUE},
if \code{TRUE} then asserts that each portfolio sums to 1 and is
non-negative. Never checks
the first portfolio, so be careful.
If \code{FALSE}, performs no checks.}
}
\value{
A matrix which has the same number of columns as
    \code{strategy$price_relatives} and one more row.
    Row \eqn{i} represents the portfolio before trading during
    period \eqn{i} (Recall that trading during period \eqn{i}
    happens before the price relatives for that period are known,
    i.e. the trade occurs, then the price changes according to the
    relatives). The column names of this matrix
    match the column names of \code{strategy$price_relatives}.
}
\description{
Run a portfolio selection strategy from the given trading
period until the end of its data, reporting the cumulative
wealth at each stage and the portfolio before each
trading period
}
