% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_utils.R
\name{compute_prices_from_relatives}
\alias{compute_prices_from_relatives}
\title{get prices from price relatives}
\usage{
compute_prices_from_relatives(price_relatives, initial_prices)
}
\arguments{
\item{price_relatives}{a matrix of price relatives
where each row represents
a trading period and each column represents an asset}

\item{initial_prices}{a vector whose \eqn{i}th entry is the price
of the asset represented by column \eqn{i}
of \code{price_relatives} immediately before trading period 1}
}
\value{
A matrix with one more row and the same
     number of columns as \code{price_relatives}
     representing the true prices of the assets.
}
\description{
Given price_relatives and the initial prices, get the
actual prices
}
