% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_data.R
\name{evaluate_momentum_predict_price}
\alias{evaluate_momentum_predict_price}
\title{Evaluate momentum price prediction}
\usage{
evaluate_momentum_predict_price(
  agg_windows,
  momentum_colname = "momentum",
  consider_negative_momentum = TRUE
)
}
\arguments{
\item{momentum_colname}{The name of the column holding the
true momentum values (as a string),
defaults to \code{"momentum"}.}

\item{consider_negative_momentum}{\code{TRUE} to treat momentum as
possibly negative, otherwise assumed to be nonnegative
(see \code{\link{evaluate_momentum_at_window}()})}

\item{aggregated_momentum_data}{A tibble
of class \code{@eval{AGG_WINDOW_CLASS_NAME}} as returned from
\code{\link{aggregate_price_and_mean_windows}}, with an
added column of column name \code{momentum_colname}
holding the momentum.}
}
\value{
A column vector whose \eqn{i}th entry is the predicted
    price for the \eqn{i}th row, using the predicted momentum
    as a guide.
}
\description{
Give the price predicted by the momentum.
}
