% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFselect.R
\name{generate_knockoff_data}
\alias{generate_knockoff_data}
\title{Generate Knockoff Data for SNPs}
\usage{
generate_knockoff_data(genetic_variants)
}
\arguments{
\item{snp_data}{Data frame containing SNP and covariate data.}
}
\value{
Knockoff genotype matrix with the same dimensions as the original genotype matrix.
}
\description{
This function takes a genotype data frame, extracts the SNP columns, and generates knockoff data.
}
