% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_buyandhold}
\alias{new_buyandhold}
\alias{new_uniform_bah}
\alias{new_best_stock}
\title{makes a \code{\link[=new_buyandhold]{buyandhold}} strategy}
\usage{
new_buyandhold(price_relatives, transaction_rate, ..., class = character())

new_uniform_bah(price_relatives, transaction_rate, ..., class = character())

new_best_stock(
  price_relatives,
  transaction_rate,
  last_trading_period,
  ...,
  class = character()
)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given \code{price_relatives}
     \code{nassets} and \code{ntrading_periods}}

\item{last_trading_period}{the last trading period after
which we "sell" our stocks (we need this to determine
which stock is the best for the \code{best_stock} strategy).}
}
\value{
a new \code{buyandhold} object with the given price relatives
}
\description{
Makes a \code{\link[=new_buyandhold]{buyandhold}} strategy from the given price
relatives, which is a subclass of \link[=new_pfselectstrat]{pfselectstrat}.
}
\details{
To validate the object use \code{\link{validate_buyandhold}}.
}
\section{Functions}{
\itemize{
\item \code{new_uniform_bah}: Starts with a uniform amount of wealth in each stock.

\item \code{new_best_stock}: Makes a \link[=new_buyandhold]{buyandhold} strategy which invests all its
wealth in the best stock from the start of the trading period
through the given last_trading_period
}}

\note{
An instance which is just of class \code{buyandhold} and on subclass
    has no first_portfolio method
}
\seealso{
new_pfselectstrat

Other buyandhold: 
\code{\link{best_stock}()},
\code{\link{uniform_bah}()}
}
\concept{buyandhold}
