% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{new_pfselectstrat}
\alias{new_pfselectstrat}
\title{makes a new \code{"pfselectstrat"} class}
\usage{
new_pfselectstrat(price_relatives, transaction_rate, ..., class = character())
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the
price after the trading period divided by the price
during the trading period.}

\item{transaction_rate}{a transaction rate in \eqn{[0,1]},
the percentage of each buy/sell lost to fees}

\item{...}{Arguments for subclassing, these are NOT set as attributes
but instead are passed as arguments into the list which
underlies this data structure}

\item{class}{A character vector of the child classes this instance
    has (or \code{character()} if no child classes)

 @return A new pfselectstrat object with the given \code{price_relatives}
     \code{nassets} and \code{ntrading_periods}}
}
\description{
Creates a \code{"pfselectstrat"} class, which has a number
of assets and a number of trading periods. We think of each
trading period as corresponding to a change of prices. At the
beginning of the trading period the prices have not yet changed,
and at the end of the trading period they have changed to the next prices.
}
\details{
The class is functionally a list with named entries
\describe{
   \item{nassets}{The number of assets}
   \item{ntrading_periods}{The number of trading periods}
   \item{price_relatives}{A \code{ntrading_periods} x \code{nassets}
   matrix of price relatives (price after trading period / price
   at beginning of trading period)}
   \item{transaction_rate}{A rate in \eqn{[0,1]} which is
   lost during each buy/sell}
}

For validation of \code{pfselectstrat} objects, see
\code{\link{validate_pfselectstrat}()}
}
\note{
This is a class with no next_portfolio and no first_portfolio
    method
}
\seealso{
Other pfselectstrat: 
\code{\link{first_portfolio}()},
\code{\link{next_portfolio}()},
\code{\link{validate_pfselectstrat}()}
}
\concept{pfselectstrat}
