% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{predict_momentum_LOAD}
\alias{predict_momentum_LOAD}
\title{Test momentum prediction using LOAD strategy}
\usage{
predict_momentum_LOAD(
  decay_factor,
  regularization_factor,
  time_window,
  momentum_threshold,
  prices
)
}
\arguments{
\item{decay_factor}{\eqn{\alpha} in the referenced paper, LOAD predicts
that stocks regressing to the mean have approximate price
\eqn{MA_t = \alpha p_t + (1-\alpha)MA_{t-1}}, \eqn{MA_1 = p_1}}

\item{regularization_factor}{\eqn{\lambda} in the referenced paper,
the regularization coeffecient for weight decay when
regressing the prices in the recent time window.
Must be non-negative}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{momentum_threshold}{\eqn{\eta} in the referenced paper.
If the regressed slope is greater than \eqn{\eta} we say the
stock has momentum. This must be greater than 0.}

\item{prices}{The prices during the trading periods (a matrix
with one more row than \code{price_relatives}).}

\item{price_means}{The mean price \eqn{MA_t} is \code{decay_factor} * \eqn{p_t}
+ \code{(1-decay_factor)} * \eqn{MA_{t-1}}, a matrix the
same size as \code{prices}.}
}
\value{
A ntime_periods x nassets matrix, with each entry
    LOAD's predicted momentum for that time period (i.e.
    before seeing that time period) or NA
}
\description{
Test LOAD's strategy of predicting momentum. Returns
each momentum prediction.
See \code{\link{evaluate_momentum}} for a description
of momentum
}
