% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentum_prediction.R
\name{predict_momentum_reg_pocket}
\alias{predict_momentum_reg_pocket}
\title{Predict the momentum using a cross-validated regularized pocket algorithm}
\usage{
predict_momentum_reg_pocket(
  prices,
  time_window,
  price_means,
  weight_elimination,
  weight_decay,
  nfolds,
  maxit_per_fold,
  num_train
)
}
\arguments{
\item{prices}{The prices during the trading periods (a matrix
with one more row than \code{price_relatives}).}

\item{time_window}{\eqn{w} in the referenced paper.
The number of trading periods (including
the most recent) to regress on (must be at least 2)}

\item{price_means}{the historic mean of the prices (exponentially
weighted average of past prices)}

\item{weight_elimination}{a vector of weight elimination constants to try}

\item{weight_decay}{a vector of weight decay constants to try}

\item{nfolds}{Number of folds, if missing defaults to folds of size 1}

\item{maxit_per_fold}{maxits per fold}
}
\description{
Cross validates over the first
num_train parameters from the given weight elimination parameters
and weight decay parameters over time. Then predicts
the remaining momenta.
}
