% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfselect_strategy.R
\name{uniform_bah}
\alias{uniform_bah}
\title{Makes a new \code{uniform_bah} strategy}
\usage{
uniform_bah(price_relatives)
}
\arguments{
\item{price_relatives}{A T x n matrix of price relatives,
where each row represents a trading period and each column
represents an asset. A price relative is the current
price divided by the previous price.}
}
\value{
a new \code{buyandhold} object with the given price relatives
}
\description{
Makes a \code{buyandhold} strategy from the given price
relatives, which is a subclass of \code{pfselectstrat}.
}
\seealso{
new_uniform_bah

Other buyandhold: 
\code{\link{best_stock}()},
\code{\link{new_buyandhold}()},
\code{\link{validate_buyandhold}()}
}
\concept{buyandhold}
